Add authorization checks of exact user in games and teams

+++ foosballv2s.WebService/Controllers/TeamController.cs
@@ -48,10 +48,18 @@ namespace foosballv2s.WebService.Controllers
         [HttpPost]
         public IActionResult Post([FromBody] Team team)
         {
+           
             if (team == null)
             {
                 return new BadRequestResult();
             }
+            
+            var user = _userManager.Users.FirstOrDefault(u => u.Email == User.Identity.GetUserId());
+            if (user.Id != team.User.Id)
+            {
+                return new UnauthorizedResult();
+            }
+            
             Team updatedTeam =_repository.Add(team);
             return new ObjectResult(updatedTeam);
         }
@@ -65,6 +73,13 @@ namespace foosballv2s.WebService.Controllers
             {
                 return BadRequest();
             }
+            
+            var user = _userManager.Users.FirstOrDefault(u => u.Email == User.Identity.GetUserId());
+            if (user.Id != team.User.Id)
+            {
+                return new UnauthorizedResult();
+            }
+            
             if (_repository.Update(id, team))
             {
                 return new NoContentResult();
@@ -77,6 +92,13 @@ namespace foosballv2s.WebService.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            var team = _repository.Get(id);
+            var user = _userManager.Users.FirstOrDefault(u => u.Email == User.Identity.GetUserId());
+            if (team != null && user.Id != team.User.Id)
+            {
+                return new UnauthorizedResult();
+            }
+            
             if (_repository.Remove(id))
             {    
                 return new NoContentResult();
